pyproject.toml
[project]
name = "burger-api"
version = "0.1.0"
description = "Backend implementation for Stellar Burgers web-app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "snapshottest",
]

[tool.mypy]
plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.uv.sources]
snapshottest = { git = "https://github.com/syrusakbary/snapshottest.git" }

[tool.pytest.ini_options]
markers = [
    "now: sets datetime value for util function in db.utils",
]
